# public/resources/swagger/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Préstamos",
        "description": "Documentación generada automáticamente con OpenAPI/Swagger para la API de préstamos.",
        "contact": {
            "name": "API prestamos",
            "email": "isaaccaliz305@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor local de desarrollo"
        }
    ],
    "paths": {
...
        "/api/clientes": {
            "get": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Obtener lista de clientes",
                "description": "Devuelve todos los clientes registrados (requiere autenticación).",
                "operationId": "f8eafbc2e9500413091c7f84a4470b84",
                "responses": {
                    "200": {
                        "description": "Lista de clientes obtenida correctamente.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Cliente"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Crear nuevo cliente",
                "description": "Registra un nuevo cliente (requiere autenticación).",
                "operationId": "2a8a28c7aec782e1af95193a65053e68",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "primer_nombre",
                                    "primer_apellido",
                                    "dpi",
                                    "nit",
                                    "fecha_nacimiento"
                                ],
                                "properties": {
                                    "primer_nombre": {
                                        "type": "string",
                                        "example": "Luis"
                                    },
                                    "segundo_nombre": {
                                        "type": "string",
                                        "example": "Fernando"
                                    },
                                    "primer_apellido": {
                                        "type": "string",
                                        "example": "Hernández"
                                    },
                                    "segundo_apellido": {
                                        "type": "string",
                                        "example": "Gómez"
                                    },
                                    "dpi": {
                                        "type": "string",
                                        "example": "1234567890101"
                                    },
                                    "nit": {
                                        "type": "string",
                                        "example": "1234567-8"
                                    },
                                    "fecha_nacimiento": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1990-05-20"
                                    },
                                    "direccion": {
                                        "type": "string",
                                        "example": "Ciudad de Guatemala, Zona 1"
                                    },
                                    "correo": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "luis@example.com"
                                    },
                                    "telefono": {
                                        "type": "string",
                                        "example": "5555-1234"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Cliente creado correctamente"
                    },
                    "422": {
                        "description": "Error de validación"
                    },
                    "401": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/clientes/{id}": {
            "get": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Mostrar cliente específico",
                "description": "Obtiene los datos de un cliente por su identificador (requiere autenticación).",
                "operationId": "7be647ff4c2ab71c5a1f042133b13428",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del cliente a recuperar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cliente encontrado"
                    },
                    "404": {
                        "description": "Cliente no encontrado"
                    },
                    "401": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Actualizar cliente",
                "description": "Modifica los datos de un cliente existente (requiere autenticación).",
                "operationId": "1cc735d14964cb4b25557d546e807a9c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cliente actualizado correctamente"
                    },
                    "404": {
                        "description": "Cliente no encontrado"
                    },
                    "401": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Eliminar cliente",
                "description": "Elimina un cliente existente y sus datos relacionados (requiere autenticación).",
                "operationId": "fae533cf0841c3ece1414287d89ef42b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Cliente eliminado correctamente"
                    },
                    "404": {
                        "description": "Cliente no encontrado"
                    },
                    "401": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Cliente": {
                "title": "Cliente",
                "description": "Modelo que representa un cliente en el sistema",
                "required": [
                    "id",
                    "primer_nombre",
                    "primer_apellido",
                    "dpi",
                    "nit"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "primer_nombre": {
                        "type": "string",
                        "example": "Juan"
                    },
                    "segundo_nombre": {
                        "type": "string",
                        "example": "Carlos"
                    },
                    "primer_apellido": {
                        "type": "string",
                        "example": "Pérez"
                    },
                    "segundo_apellido": {
                        "type": "string",
                        "example": "López"
                    },
                    "dpi": {
                        "type": "string",
                        "example": "1234567890101"
                    },
                    "nit": {
                        "type": "string",
                        "example": "1234567-8"
                    },
                    "fecha_nacimiento": {
                        "type": "string",
                        "format": "date",
                        "example": "1990-05-10"
                    },
                    "direccion": {
                        "type": "string",
                        "example": "Ciudad de Guatemala, Zona 1"
                    },
                    "correo": {
                        "type": "string",
                        "format": "email",
                        "example": "juan.perez@example.com"
                    },
                    "telefono": {
                        "type": "string",
                        "example": "5555-1234"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}